+++ CHANGES.md
@@ -1,5 +1,17 @@
 # Changelog
 
+**8.3.3** (2023-05-12)
+  * Fixed Python 3.8 compat
+
+**8.3.2** (2023-05-11)
+  * Updated docs about permission settings setup
+
+**8.3.1** (2023-05-11)
+  * Updated docs about permission feature limitations
+
+**8.3.0** (2023-05-11)
+  * Added improved Django group permission fixtures
+
 **8.2.1** (2023-05-10)
   * Updated package linters
   * Added release scripts for windows and UNIX